+++ example1/lib/src/util/examples_descriptor.dart
@@ -38,7 +38,8 @@ enum ExamplesEnum {
   ex11RandomDataWithLabelLayoutStrategy,
   ex30AnimalsBySeasonWithLabelLayoutStrategy,
   ex31SomeNegativeValues,
-  ex32SmallPositivesAxisStartsAbove0,
+  ex32AllPositiveYsYAxisStartsAbove0,
+  ex33AllNegativeYsYAxisEndsBelow0,
   ex35AnimalsBySeasonNoLabelsShown,
   ex40LanguagesWithYOrdinalUserLabelsAndUserColors,
   ex50StocksWithNegativesWithUserColors,
@@ -81,8 +82,10 @@ class ExamplesDescriptor {
     const Tuple2(ExamplesEnum.ex31SomeNegativeValues, ExamplesChartTypeEnum.lineChart),
     const Tuple2(ExamplesEnum.ex31SomeNegativeValues, ExamplesChartTypeEnum.verticalBarChart),
     //
-    const Tuple2(ExamplesEnum.ex32SmallPositivesAxisStartsAbove0, ExamplesChartTypeEnum.lineChart),
-    const Tuple2(ExamplesEnum.ex32SmallPositivesAxisStartsAbove0, ExamplesChartTypeEnum.verticalBarChart),
+    const Tuple2(ExamplesEnum.ex32AllPositiveYsYAxisStartsAbove0, ExamplesChartTypeEnum.lineChart),
+    const Tuple2(ExamplesEnum.ex32AllPositiveYsYAxisStartsAbove0, ExamplesChartTypeEnum.verticalBarChart),
+    //
+    const Tuple2(ExamplesEnum.ex33AllNegativeYsYAxisEndsBelow0, ExamplesChartTypeEnum.lineChart),
     //
     const Tuple2(ExamplesEnum.ex35AnimalsBySeasonNoLabelsShown, ExamplesChartTypeEnum.lineChart),
     const Tuple2(ExamplesEnum.ex35AnimalsBySeasonNoLabelsShown, ExamplesChartTypeEnum.verticalBarChart),